fix warnings && remove useless file

+++ srcs/fdf/scale.c
@@ -62,7 +62,7 @@ static float	find_highest(t_vec3 *pts, t_size map_size)
 	float	res;
 
 	idx = (size_t)-1;
-	end = map_size.width * map_size.height;
+	end = (size_t)map_size.width * (size_t)map_size.height;
 	res = 0.0f;
 	while (++idx < end)
 		if (ft_absf(pts[idx].y) > res)
@@ -75,7 +75,7 @@ void	fdf_scale_to_fit(t_fdf *fdf)
 	float	section_scale;
 	float	highest;
 
-	section_scale = fdf->map_size.width;
+	section_scale = (float)fdf->map_size.width;
 	highest = find_highest(fdf->orig_points, fdf->map_size);
 	fdf->scale = (t_vec3){1 / section_scale, 1 / highest, 1 / section_scale};
 }
